+++ src/components/borderBox11/index.js
@@ -10,12 +10,14 @@ import { fade } from '@jiaminghi/color'
 
 import useAutoResize from '../../use/autoResize'
 
+import { uuid } from '../../util'
+
 import './style.less'
 
 const defaultColor = ['#8aaafb', '#1f33a2']
 
 const BorderBox = forwardRef(({ children, className, style, color = [], titleWidth = 250, title = '', backgroundColor = 'transparent' }, ref) => {
-  const filterId = useRef(`border-box-11-filterId-${Date.now()}`).current
+  const filterId = useRef(`border-box-11-filterId-${uuid()}`).current
 
   const { width, height, domRef } = useAutoResize(ref)