+++ src/components/capsuleChart/index.js
@@ -39,16 +39,20 @@ const defaultConfig = {
    * @default unit = ''
    */
   unit: '',
-  showVal: false
+  /**
+   * @description Show item value
+   * @type {Boolean}
+   * @default showValue = false
+   */
+  showValue: false
 }
 
 const CapsuleChart = ({ config = {}, className, style }) => {
-  const [{ mergedConfig, labelData, capsuleLength, capsuleValue, labelDataLength }, setState] = useState({
+  const [{ mergedConfig, labelData, capsuleLength, capsuleValue }, setState] = useState({
     mergedConfig: null,
     labelData: [],
     capsuleLength: [],
-    capsuleValue: [],
-    labelDataLength: []
+    capsuleValue: []
   })
 
   useEffect(() => {
@@ -62,31 +66,15 @@ const CapsuleChart = ({ config = {}, className, style }) => {
 
     const maxValue = Math.max(...capsuleValue)
 
-    const capsuleLength = capsuleValue.map(v => (maxValue ? v / maxValue : 0))
-
     const oneFifth = maxValue / 5
-
-    const labelData = [...new Set(new Array(6).fill(0).map((v, i) => Math.ceil(i * oneFifth)))]
-
-    const labelDataLength = labelData.map(v => maxValue ? v / maxValue : 0)
 
     setState({
       mergedConfig,
       capsuleValue,
-      capsuleLength,
-      labelData,
-      labelDataLength
+      capsuleLength: capsuleValue.map(v => (maxValue ? v / maxValue : 0)),
+      labelData: [...new Set(new Array(6).fill(0).map((v, i) => Math.ceil(i * oneFifth)))]
     })
   }, [config])
-
-  /**
-   * 计算x轴label位置
-   */
-  function calcUnitLabelStyle(index) {
-    return labelData.length - 1 === index
-      ? `right: 0;`
-      : `left: ${labelDataLength[index] * 100}%;`
-  }
 
   const classNames = useMemo(() => classnames('dv-capsule-chart', className), [
     className
@@ -107,6 +95,7 @@ const CapsuleChart = ({ config = {}, className, style }) => {
             {capsuleLength.map((capsule, index) => (
               <div className='capsule-item' key={index}>
                 <div
+                  className='capsule-item-column'
                   style={{
                     width: `${capsule * 100}%`,
                     backgroundColor: `${
@@ -115,19 +104,18 @@ const CapsuleChart = ({ config = {}, className, style }) => {
                   }}
                 >
                   {
-                    mergedConfig.showVal &&
-                      <span
-                        className='capsule-item-val'
-                        style={`right:-${`${capsuleValue[index]}`.length * 7}px`}>
+                    mergedConfig.showValue &&
+                      <div className='capsule-item-value'>
                         { capsuleValue[index] }
-                      </span>
+                      </div>
                   }
                 </div>
               </div>
             ))}
+
             <div className='unit-label'>
               {labelData.map((label, index) => (
-                <div key={label + index} style={calcUnitLabelStyle(index)}>{label}</div>
+                <div key={label + index}>{label}</div>
               ))}
             </div>
           </div>